+++ handlers/actions/set-bot-commands.ts
@@ -2,4 +2,4 @@ import { bot } from "../..";
 
 // TODO: set custom commands here instead of @botfather
 // https://core.telegram.org/bots/api#setmycommands
-await bot.api.setMyCommands([]);
+bot.api.setMyCommands([]);
