+++ handlers/commands/index.ts
@@ -1,5 +1,4 @@
-// import "./admin";
+import "./admin";
 import "./autoexpand";
 // import "./news";
-// import "./source-code";
-import "./start";
+import "./source-code";
