--- handlers/commands/start.ts
@@ -1,5 +1,0 @@
-import { bot } from "../..";
-
-bot.command("start", async (ctx) => {
-  ctx.reply("welcome");
-});
